requirements.txt
py-cord
PyNaCl
pynacl
pygame
python-dotenv
yt-dlp
discord.py
spotify
spotipy
Flaskk